chore: Consolidate IO-related code into single module
The changes show merging fn_utils.rs into io.rs and organizing the code with clear section headers. The IO-related functionality is now
better organized in a single module with improved documentation.
