+++ src/data/mod.rs
@@ -1,5 +1,4 @@
-pub mod fn_utils;
-pub mod io_utils;
+pub mod io;
 pub mod stores;
 pub mod themes;
 pub mod types;
